Original Equation Code
This is the Original Code of the research paper published in this repository. This code have all the conditions needed to solve the Tensor original. (∇^2 * h = 0). All the code is commented in Spanish, so we can know what is the function of every line and help us to understand how the numerical method is implemented to solve the differential equation (tensor) in Matlab.