Modified Equation Code
This is the Code worked as an exercise to understand the syntax of Matlab, implementing new frontier conditions given by the change of the equation. With these new conditions reformulated and given we already can solve the Tensor. (∇^2 * h = e^(-(x^2+y^2))). All the code is commented in Spanish, so we can know what is the function of every line and help us to understand how the numerical method is implemented to solve the differential equation (tensor) in Matlab.